committing the rest of vscode auto refactor
