+++ openoa/analysis/plant_analysis.py
@@ -16,14 +16,14 @@ import statsmodels.api as sm
 from tqdm import tqdm
 from openoa import logging, logged_method_call
 from openoa.types import timeseries_table
-from openoa.toolkits import filters
-from openoa.toolkits import timeseries as tm
-from openoa.toolkits import unit_conversion as un
-from openoa.toolkits import met_data_processing as mt
+from openoa.utils import filters
+from openoa.utils import timeseries as tm
+from openoa.utils import unit_conversion as un
+from openoa.utils import met_data_processing as mt
 from sklearn.metrics import r2_score, mean_squared_error
 from sklearn.linear_model import LinearRegression
 from sklearn.model_selection import KFold
-from openoa.toolkits.machine_learning_setup import MachineLearningSetup
+from openoa.utils.machine_learning_setup import MachineLearningSetup
 
 
 logger = logging.getLogger(__name__)
