+++ test/regression/int_electrical_losses.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import numpy as np
 import numpy.testing as npt
 from examples.project_ENGIE import Project_Engie
-from openoa.methods.electrical_losses import ElectricalLosses
+from openoa.analysis.electrical_losses import ElectricalLosses
 
 
 class TestElectricalLosses(unittest.TestCase):
