+++ test/regression/int_turbine_long_term_gross_energy.py
@@ -5,7 +5,7 @@ from test import example_data_path_str
 import numpy as np
 import numpy.testing as npt
 from examples.project_ENGIE import Project_Engie
-from openoa.methods.turbine_long_term_gross_energy import TurbineLongTermGrossEnergy
+from openoa.analysis.turbine_long_term_gross_energy import TurbineLongTermGrossEnergy
 
 
 def reset_prng():
