+++ test/unit/test_ml_toolkit.py
@@ -6,7 +6,7 @@ import pandas as pd
 from numpy import testing as nptest
 from sklearn.metrics import mean_squared_error
 from sklearn.model_selection import KFold
-from openoa.toolkits.machine_learning_setup import MachineLearningSetup
+from openoa.utils.machine_learning_setup import MachineLearningSetup
 
 
 class TestMLToolkit(unittest.TestCase):
