+++ test/unit/test_power_curve_toolkit.py
@@ -3,8 +3,8 @@ import unittest
 import numpy as np
 import pandas as pd
 from numpy import testing as nptest
-from openoa.toolkits import power_curve
-from openoa.toolkits.power_curve.parametric_forms import logistic5param, logistic5param_capped
+from openoa.utils import power_curve
+from openoa.utils.power_curve.parametric_forms import logistic5param, logistic5param_capped
 
 
 noise = 0.1
