+++ test/unit/test_timeseries_toolkit.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from pytz import timezone
 from numpy import testing as nptest
-from openoa.toolkits import timeseries
+from openoa.utils import timeseries
 
 
 class SimpleTimeseriesTests(unittest.TestCase):
